requirements.txt
fastapi
uvicorn
python-dotenv
fdb
unidecode
